requirements.txt
pandas==2.2.2
pyarrow==16.1.0
kafka-python==2.0.2
confluent-kafka==2.6.0
python-dotenv==1.0.1
pydantic==2.9.2
jsonschema==4.23.0
pyyaml==6.0.2
